Compiling again - Dropped the whole Heap memory arena. Its to limiting at this point in the project and I don't want to deal with it. Not enough time

+++ PhxEngine/RHI/RHITypes.h
@@ -637,7 +637,7 @@ namespace phx::rhi
     using ShaderModuleHandle = Handle<ShaderModule>;
     struct ShaderModuleDescriptor
     {
-        phx::Span<uint8_t> byte_code;
+        phx::Span<uint32_t> byte_code;
         bool IsValid() const { return !byte_code.IsEmpty(); }
     };
